pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
asgiref = "^3.4.1"
autopep8 = "^1.5.7"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
pycodestyle = "^2.7.0"
pytz = "^2021.3"
sqlparse = "^0.4.2"
toml = "^0.10.2"
Unipath = "^1.1"
whitenoise = "^5.3.0"
python-decouple = "^3.5"
javascript = "^2015.32.59"
djangorestframework = "^3.12.4"
[tool.poetry.dev-dependencies]
